+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - stricter type checking
 - add prettier
 - format codebase
+- replace Hono with URLpattern
 
 ## [0.4.1] - 2025-10-24